Fix install.sh script to work from current repo
The install.sh script previously assumed that the repo would always be
checked out at /home/$user.
So, if we ran the script when the repo was checked out at another location,
the script would still go to /home/$user and use the wrong repo.

This fix changes the script to work from the repo that contains the
install.sh script that was invoked.